pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "contact-manager"
version = "0.1.0"
description = "Contact Manager System with login feature"
authors = ["Diego Godoi <diegogodoimartins.11@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
aiosqlite = "^0.22.1"
bcrypt = "^5.0.0"
fastapi = { extras = ["standard"], version = "^0.135.2" }
pydantic = "^2.12.5"
python = ">=3.14,<4.0"
python-jose = "^3.5.0"
python-multipart = "^0.0.22"
slowapi = "^0.1.9"
sqlalchemy = "^2.0.48"
uvicorn = "^0.42.0"

[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.3"
faker = "^40.11.1"
httpx = "^0.28.1"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.1.0"
pytest-mock = "^3.15.1"
ruff = "^0.15.7"
taskipy = "^1.12.0"

[tool.ruff]
extend-exclude = ["migrations"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR6301", "PLR2004"]

[tool.ruff.format]
preview = true
quote-style = "single"

[tool.taskipy.tasks]
clean_test = "rm -rf .pytest_cache .coverage htmlcov && task test"
clear = "rm -rf .pytest_cache .coverage htmlcov"
format = "ruff format"
lint = "ruff check"
post_test = "coverage html"
pre_format = "ruff check --fix"
pre_test = "task lint"
run = "fastapi dev run.py --proxy-headers --forwarded-allow-ips='*'"
test = "pytest -s -x --cov=app tests/ -v"
